Add option to inherit options from parent sites
TODO: I don't like the UI, because for it to really be useful it should
sync stuff from different boxes (e.g. the localisation settings, IP)
which I don't really like. I also just realized that the "user" field is
always linked to the parent site (which is okay), but that's not very
clear in the UI.